refactor: 리스트 API 경로 수정

+++ src/main/java/com/example/panttegi/list/service/ListService.java
@@ -9,6 +9,8 @@ import com.example.panttegi.list.entity.BoardList;
 import com.example.panttegi.list.repository.ListRepository;
 import com.example.panttegi.user.entity.User;
 import com.example.panttegi.user.repository.UserRepository;
+import com.example.panttegi.workspace.entity.Workspace;
+import com.example.panttegi.workspace.repository.WorkspaceRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -23,11 +25,17 @@ public class ListService {
     private final ListRepository listRepository;
     private final BoardRepository boardRepository;
     private final UserRepository userRepository;
+    private final WorkspaceRepository workspaceRepository;
 
     @Transactional
-    public ListResponseDto createList(Long boardId, String title, String email) {
+    public ListResponseDto createList(Long workspaceId, Long boardId, String title, String email) {
         User user = userRepository.findByEmailOrElseThrow(email);
+        Workspace workspace = workspaceRepository.findByIdOrElseThrow(workspaceId);
         Board board = boardRepository.findByIdOrElseThrow(boardId);
+
+        if (!board.getWorkspace().getId().equals(workspace.getId())) {
+            throw new CustomException(ErrorCode.BAD_REQUEST);
+        }
 
         List<BoardList> lists = listRepository.findByBoardId(boardId);
         lists.sort(Comparator.comparing(BoardList::getPosition));
@@ -39,9 +47,12 @@ public class ListService {
     }
 
     @Transactional
-    public ListResponseDto updateList(Long listId, String title, int targetIndex, String email) {
-        User user = userRepository.findByEmailOrElseThrow(email);
+    public ListResponseDto updateList(Long workspaceId, Long listId, String title, int targetIndex, String email) {
         BoardList targetList = listRepository.findByIdOrElseThrow(listId);
+
+        if (!targetList.getBoard().getWorkspace().getId().equals(workspaceId)) {
+            throw new CustomException(ErrorCode.BAD_REQUEST);
+        }
 
         List<BoardList> lists = listRepository.findByBoardId(targetList.getBoard().getId());
         lists.sort(Comparator.comparing(BoardList::getPosition));
@@ -61,8 +72,13 @@ public class ListService {
     }
 
     @Transactional
-    public void deleteList(Long listId, String email) {
+    public void deleteList(Long workspaceId, Long listId, String email) {
         BoardList boardList = listRepository.findByIdOrElseThrow(listId);
+
+        if (!boardList.getBoard().getWorkspace().getId().equals(workspaceId)) {
+            throw new CustomException(ErrorCode.BAD_REQUEST);
+        }
+
         listRepository.delete(boardList);
     }
